Add files via upload
The dataset being used in this project was obtained from Kaggle.com and contains data from the housing market in Nashville, Tennessee. In this project, I construct an array of queries to effectively clean the data and ensure it's accuracy including removing duplicates, breaking important data into distinct columns, and more.